Created average and best chromosome summary logs

+++ SolvitaireGenetics/IO/AgentLog.cs
@@ -8,7 +8,7 @@ namespace SolvitaireGenetics;
 public class AgentLog
 {
     public int Generation { get; set; }
-    public int Count { get; set; } = 1;
+    public float Count { get; set; } = 1;
     public double Fitness { get; set; }
     public int GamesWon { get; set; }
     public int MovesMade { get; set; }
